chore: allow import template inside the page
